Refactor: Extract form modal from CreateEventModal
