+++ src/EventTickets/resources/admin/components/EventFormModal/EventFormModal.module.scss
@@ -1,5 +1,5 @@
-:global(#give-admin-event-tickets-root) {
-    .createEventButton {
+.createEventForm {
+    .submitButton {
         border-radius: 0;
         font-size: 0.875rem;
         font-weight: 600;
@@ -8,3 +8,4 @@
         text-align: center;
     }
 }
+
